contact page full ajax

+++ DJANGO/fromzerotohero/panel/views.py
@@ -5,8 +5,8 @@ from subcategory.models import SubCategory
 from django.core.files.storage import FileSystemStorage
 from main.models import Main, ContactForm
 from django.http import JsonResponse
-from django.core.serializers import serialize
-
+from django.core import serializers
+import json
 
 # Create your views here.
 def home(request):
@@ -423,53 +423,43 @@ def message_box(request):
     if request.method == "POST" and request.is_ajax():
         if request.POST.get('process') and request.POST.get('process') == "is_it_read":
             pk = request.POST.get('message_pk')
-            message = ContactForm.objects.get(pk=pk)
-            message.is_it_read = True
-            message.save()
+            try:
+                message = ContactForm.objects.get(pk=pk)
+                message.is_it_read = True
+                message.save()
+                data = {
+                    'success': True,
+                    'name': message.name,
+                    'email': message.email,
+                    'message': message.message,
+                }
+            except:
+                data = {
+                    'success': False,
+                }
+            return JsonResponse(data)
+        elif request.POST.get('process') and request.POST.get('process') == 'get_page_data':
+            messages = serializers.serialize('json', ContactForm.objects.all())
             data = {
+                'messages': messages,
                 'success': True,
             }
             return JsonResponse(data)
-        # elif request.POST.get('process') and request.POST.get('process') == 'get_page_data':
-        #     messages = serialize('json', ContactForm.objects.all())
-        #     data = {
-        #         'messages': messages,
-        #         'success': True,
-        #     }
-        #     return JsonResponse(data)
+        elif request.POST.get('process') and request.POST.get('process') == 'delete_message':
+            pk = request.POST.get('message_pk')
+            try:
+                message = ContactForm.objects.get(pk=pk)
+                message.delete()
+                data = {
+                    'success': True,
+                }
+            except:
+                data = {
+                    'success': False,
+                }
+            return JsonResponse(data)
     messages = ContactForm.objects.all()
     context = {
         'messages': messages,
     }
     return render(request, 'back/message_box.html', context=context)
-
-
-def message_box_delete(request, pk):
-    try:
-        message = get_object_or_404(ContactForm, pk=pk)
-        message.delete()
-        return redirect('panel:message_box')
-    except:
-        error = "Something Wrong!"
-        link = request.META.get('HTTP_REFERER')
-        context = {
-            'error': error,
-            'link': link
-        }
-        return render(request, 'back/error.html', context=context)
-
-
-def message_box_update(request, pk):
-    try:
-        message = get_object_or_404(ContactForm, pk=pk)
-        message.is_it_read = True
-        message.save()
-        return redirect('panel:message_box')
-    except:
-        error = "Something Wrong!"
-        link = request.META.get('HTTP_REFERER')
-        context = {
-            'error': error,
-            'link': link
-        }
-        return render(request, 'back/error.html', context=context)
